refactor: 팀 참여하기 섹션 분리(#20)
